Add files via upload
The basic commands. Support movement via the left joystick as well as the D-Pad, access to the map (Square), reporting of bodies (Circle), Vent/Use (RZ/R) and Killing (LZ/L).
Also the menu can be called via the Start (or possibly Select) Button, though the coordinates here may be wrong, you probably will need to change that.

This program uses only the two libraries pynput and inputs [and the dependencies], both of which can be installed via pip. (pip install pynput inputs)